category & sub catgory in some changes as per requriement

+++ src/category/subcategory/subcategory.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, HttpException, NotFoundException, HttpStatus } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { Subcategory } from '../category.schema/sub-category.schema';
 import { CreateSubcategoryDTO, UpdateSubcategoryDTO } from '../dto/create-subcategory.dto';
 import CustomResponse from 'src/common/providers/custom-response.service';
@@ -13,52 +13,39 @@ export class SubcategoryService {
     @InjectModel('Subcategory') private subcategoryModel: Model<Subcategory>,
     @InjectModel('Category') private categoryModel: Model<Category>, // Inject Category Model
   ) { }
-
-  private generateImageUrls(files: { app_image?: Express.Multer.File[]; web_image?: Express.Multer.File[] }, serverUrl: string) {
-  //   const fileName = fileUpload(ENTITY.USER, file);
-  //   updateUserDto['filePath'] =
-  //     process.env.SERVER_BASE_URL + `uploads/${ENTITY.USER}/${fileName}`;
-  // }
-    const appImageUrls = files.app_image
-      ? files.app_image.map(file => `${serverUrl}/uploads/${file.filename}`)
-      : [];
-    const webImageUrls = files.web_image
-      ? files.web_image.map(file => `${serverUrl}/uploads/${file.filename}`)
-      : [];
-
-    return { appImageUrls, webImageUrls };
-  }
-
   // Create Subcategory
-  async createSubcategory(createSubcategoryDto: CreateSubcategoryDTO, files: { app_image?: Express.Multer.File[]; web_image?: Express.Multer.File[] }) {
+  async createSubcategory(createSubcategoryDto: CreateSubcategoryDTO, files: any) {
     try {
-      const serverUrl = 'http://147.93.103.99:8000/'; // Adjust for production URL
 
-      // Check if Category exists
+      if (!createSubcategoryDto || !files) {
+        throw new NotFoundException(400, 'Missing required fields or files');
+      }
+      const webImageFile = files.find(file => file.fieldname === 'web_image');
+      const appImageFile = files.find(file => file.fieldname === 'app_image');
+
+      // Upload files using your fileUpload function
+      const webImage = fileUpload('subcategory/webImage', webImageFile || null);
+      createSubcategoryDto['web_image'] = webImage
+        ? [`${process.env.SERVER_BASE_URL}uploads/subcategory/webImage/${webImage}`]
+        : [];
+      const appImage = fileUpload('subcategory/appImage', appImageFile || null);
+      createSubcategoryDto['app_image'] = appImage
+        ? [`${process.env.SERVER_BASE_URL}uploads/subcategory/appImage/${appImage}`]
+        : [];
       const category = await this.categoryModel.findById(createSubcategoryDto.categoryId);
       if (!category) {
         throw new NotFoundException('Category not found');
       }
-      const categoryBySlug = await this.categoryModel.findOne({ slug: createSubcategoryDto.slug });
-      if (!categoryBySlug) {
-        throw new NotFoundException('Category with the specified slug not found');
-      }
-      //  console.log('files',files)
-      // Generate image URLs
-      const { appImageUrls, webImageUrls } = this.generateImageUrls(files, serverUrl);
-
-      // Add these URLs to DTO before saving
-      createSubcategoryDto.app_image = appImageUrls;
-      createSubcategoryDto.web_image = webImageUrls;
-     
-      // Create the subcategory
-      // const subcategory = new this.subcategoryModel(createSubcategoryDto);
-      // console.log('hiii',subcategory)
-      // const subSave = await subcategory.save();
+      // const categoryBySlug = await this.categoryModel.findOne({ slug: createSubcategoryDto.slug });
+      // if (!categoryBySlug) {
+      //   throw new NotFoundException('Category with the specified slug not found');
+      // }
+      const subcategory = new this.subcategoryModel(createSubcategoryDto);
+      const subSave = await subcategory.save();
       return new CustomResponse(
         HttpStatus.OK,
         'CreateSubcategory Save SuccessFully',
-        'kk'
+        subSave
       );
 
     } catch (error) {
@@ -67,21 +54,25 @@ export class SubcategoryService {
   }
 
   // Update Subcategory
-  async updateSubcategory(id: string, updateSubcategoryDto: UpdateSubcategoryDTO, files: { app_image?: Express.Multer.File[]; web_image?: Express.Multer.File[] }) {
+  async updateSubcategory(id: string, updateSubcategoryDto: UpdateSubcategoryDTO, files: any) {
     try {
-      const serverUrl = 'http://147.93.103.99:8000/';; // Adjust for production URL
-
-      // Check if Category exists
+      if (!id || !updateSubcategoryDto || !files) {
+        throw new NotFoundException(400, 'Missing required fields or files');
+      }
+      const webImageFile = files.find(file => file.fieldname === 'web_image');
+      const appImageFile = files.find(file => file.fieldname === 'app_image');
+      const webImage = fileUpload('subcategory/webImage', webImageFile || null);
+      updateSubcategoryDto['web_image'] = webImage
+        ? [`${process.env.SERVER_BASE_URL}uploads/subcategory/webImage/${webImage}`]
+        : [];
+      const appImage = fileUpload('subcategory/appImage', appImageFile || null);
+      updateSubcategoryDto['app_image'] = appImage
+        ? [`${process.env.SERVER_BASE_URL}uploads/subcategory/appImage/${appImage}`]
+        : [];
       const category = await this.categoryModel.findById(updateSubcategoryDto.categoryId);
       if (!category) {
         throw new NotFoundException('Category not found');
       }
-
-      const { appImageUrls, webImageUrls } = this.generateImageUrls(files, serverUrl);
-
-      updateSubcategoryDto.app_image = appImageUrls;
-      updateSubcategoryDto.web_image = webImageUrls;
-
       const subcategory = await this.subcategoryModel.findByIdAndUpdate(id, updateSubcategoryDto, { new: true });
       if (!subcategory) {
         throw new NotFoundException('SubCategory  not found');
@@ -100,13 +91,14 @@ export class SubcategoryService {
   // Delete Subcategory
   async deleteSubcategory(id: string) {
     try {
-      const subcategory = await this.subcategoryModel.findByIdAndDelete(id);
-      if (!subcategory) {
-        throw new NotFoundException('Category not found');
+      const subCatCheck = await this.subcategoryModel.findById(id);
+      if (!subCatCheck) {
+        throw new NotFoundException('Sub Category not found');
       }
+      const subcategory = await this.subcategoryModel.findByIdAndDelete(id);
       return new CustomResponse(
         HttpStatus.OK,
-        'CreateSubcategory Delete SuccessFully',
+        'Subcategory Delete SuccessFully',
         subcategory
       );
     } catch (error) {
@@ -152,7 +144,6 @@ export class SubcategoryService {
       // }
 
       const subcategories = await this.subcategoryModel.find({ categoryId: categoryId });
-      console.log(subcategories)
       return new CustomResponse(
         HttpStatus.OK,
         'Subcategory By Id  SuccessFully',
@@ -173,8 +164,7 @@ export class SubcategoryService {
         );
       }
 
-      const subcategories = await this.subcategoryModel.findOne({_id:id});
-      console.log('subcategory',subcategories)
+      const subcategories = await this.subcategoryModel.findOne({ _id: id });
       return new CustomResponse(
         HttpStatus.OK,
         'All Subcategories get With Category Id',
@@ -184,5 +174,59 @@ export class SubcategoryService {
       throwException(error)
     }
   }
+    async getFilteredSubCategories(filters: Record<string, any>) {
+      try {
+        const query: any = {};
+        // if (filters.id) {
+        //   // Convert 'id' to ObjectId only if it is a valid ObjectId string
+        //   if (Types.ObjectId.isValid(filters.id)) {
+        //     query._id = filters.id; // Only if the id is valid
+        //   } else {
+        //     throw new Error('Invalid ObjectId format');
+        //   }
+        // }
+        // Loop through each filter key and dynamically build the query
+        for (const [key, value] of Object.entries(filters)) {
+  
+  
+          if (value) {
+            // For boolean filters like is_published, public
+            if (key === 'is_published' || key === 'categoryId') {
+              query[key] = value;
+            }
+            // For ObjectId based filter (e.g., filtering by _id)
+            else if (key === 'id' && value) {
+              // Convert _id to ObjectId for comparison
+              query._id = new Types.ObjectId(value); // Convert string to ObjectId
+            }
+            // For date range filter (assuming filters contains date keys like 'startDate' and 'endDate')
+            else if (key === 'create_at' && value) {
+              query.createdAt = { $gte: new Date(value) }; // Greater than or equal to start date
+            } else if (key === 'update_at' && value) {
+              query.createdAt = { $lte: new Date(value) }; // Less than or equal to end date
+            }
+            // For text-based search like name, description, etc.
+            else {
+              query[key] = { $regex: value, $options: 'i' }; // Case-insensitive match
+            }
+          }
+        }
+        // console.log('Filtered Categories:', query);
+        const filterQuery = await this.subcategoryModel.find(query).exec();
+  
+  
+        // Check if no data found, return custom response
+        if (filterQuery.length === 0) {
+          return new CustomResponse(404, 'Sub Category Not Found');
+        }
+  
+        // Return categories if found
+        return new CustomResponse(200, 'Sub Categories Retrieved Successfully', filterQuery);
+  
+      } catch (error) {
+        // Handle and throw the error
+        throwException(error);
+      }
+    }
 }
